requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
astroid==1.6.6            # via pylint
backports.functools-lru-cache==1.6.1  # via astroid, isort, pylint
configparser==4.0.2       # via pylint
coverage==4.5.1           # via -r requirements/test.in
enum34==1.1.10            # via astroid
funcsigs==1.0.2           # via mock
futures==3.3.0 ; python_version == "2.7"  # via -c requirements/constraints.txt, isort
isort==4.3.21             # via pylint
lazy-object-proxy==1.5.1  # via astroid
mako==1.0.2               # via -r requirements/test.in
markupsafe==1.1.1         # via mako
mccabe==0.6.1             # via pylint
mock==2.0.0               # via -r requirements/test.in
path.py==8.2.1            # via -r requirements/test.in
pbr==5.5.0                # via mock
pycodestyle==2.5.0        # via -c requirements/constraints.txt, -r requirements/test.in
pylint==1.9.3             # via -c requirements/constraints.txt, -r requirements/test.in
pytz==2020.1              # via -r requirements/base.txt, django
singledispatch==3.4.0.3   # via astroid, pylint
six==1.15.0               # via astroid, mock, pylint, singledispatch
wrapt==1.12.1             # via astroid
